Separate get and post unit tests
